server: extend the interface to support the 'new style' negotiation
The 'new style' negotiation has an options sending phase, where one of
the options is to query the list of available exports. Therefore the
'negotiate_begin' includes an optional list of exports to advertise to
the client. The client chooses an export by name and the second phase
allows the server to either accept or reject (by closing the connection)

Signed-off-by: David Scott <dave.scott@citrix.com>