CP-22631: TLS support (libs and server only)
This adds support for the TLS-related parts of the NBD protocol.

The server runs in FORCEDTLS mode by default, but can be configured
to use NOTLS mode.

A new command-line option specifies the path to the file containing the
TLS certificate, and the NBD server uses it (unless the --no-tls option
is given).

There is also a CLI option to override the default ciphersuite.

When the server is configured to use TLS, it rejects connections
from any client that does not declare FIXED_NEWSTYLE support.
(This is as specified in the NBD protcol.)

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>