NBD server: use Lwt.finalize to propagate exceptions up
Previously we swallowed these exceptions with Lwt.try_bind, but it is
useful to report them to the user. This commit also makes the cleanup
functions more robust, because now we do not rely on carefully avoiding
double-closing the file descriptors, we just ignore any exceptions that
happen in the cleanup functions.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>
(cherry picked from commit 1ef61435844b3abb0332abd88c55743863565030)