jbuild: get flags for ppx preprocessors using ocamlfind & grep
We could actually get the library paths using "ocamlfind query", as we
did previously, but it would require much more knowledge about the
dependencies and command line options, we would have to declare these in
the jbuild file.

This commit might break the build of nbd under Windows, if it currently
compiles on it.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>