Fix coverage build using jbuilder (and old ppx_bisect)
Until https://github.com/janestreet/jbuilder/issues/57 is resolved we
can still make coverage builds by adding the necessary flags by hand, as
we are doing for the "old style" ppx derivers.

We watch the value of the `COVERAGE` environment variable. If this is
defined and set to "1", we inject the necessary `ppx_bisect` flags and
dependencies in the generated jbuild file.

The boilerplate necessary is generic and can be used with minimal or no
changes in most of our other packages.

Once the necessary changes are landed in jbuilder and ppx_bisect, we can
update the project simply updating the jbuild files only.

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>