Add Nbd_lwt_unix.with_block
The first of several functions of this kind to make it easier to write
safe code with guaranteed clean-up.  Here we ensure that Block.connect
is followed by a corresponding disconnect.

  Block.connect ...
  >>= fun b ->
  Lwt.finalize
    (fun () -> ... )
    (fun () -> Block.disconnect b)

This Block module comes from the mirage-block-unix package so I added
this to the nbd-lwt-unix.opam and lwt/jbuild files.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>