Server: don't disconnect when a write errors out
The NBD protocol says about NBD_CMD_WRITE:
"A write request. Length and offset define the location and amount of
data to be written. The client MUST follow the request header with
length number of bytes to be written to the device.
The server MUST write the data to disk, and then send the reply
message. The server MAY send the reply message before the data has
reached permanent storage.
If an error occurs, the server MUST set the appropriate error code
in the error field."

And about disconnecting during the transmission phase:
"Either side MAY initiate a hard disconnect if it detects
a violation by the other party of a mandatory condition
within this document.",
and:
"The client and the server MUST NOT initiate any form
of disconnect other than in one of the above circumstances."

And about NBD_FLAG_READ_ONLY:
"The server MAY set this flag to indicate
to the client that the export is read-only (exports might be read-only
in a manner undetectable to the server, for instance because of
permissions). If this flag is set, the server MUST error subsequent
write operations to the export."

Thus it seems that the server should not disconnect in case of
NBD_CMD_WRITE errors, but should continue processing the client's
requests.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>