Server: add support for read-only exports
Currently if read_only is false, but the block device is read-only, the
function will not fail, but an error will be logged, and write attempts
from the client will result in an EPERM error.

I've set the read_only parameter to true by default for safety.  This
will not have a significant impact on existing users of the library,
because write access did not work until recently.  The nbd-tool
command-line interface will still serve disks as read-write.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>