Add exportname option to CLI and interop tests
Add an option to specify an export name when serving a file. If
specified, only this export name will be accepted, and the client will
be allowed to list the exports.

If unspecified, the behavior is the same as before: listing the exports
is not allowed, and any export name is accepted.

I've added compatibility tests with qemu-img and nbd-client. By default,
these test are skipped if the required programs aren't installed. If the
strict option is enabled, the tests will fail instead of being skipped.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>