Correctly marshal & unmarshal requests
In the request message, accorrding to the NBD protocol, there is a
16-bit flags and a 16-bit type field But we had a single 32-bit type
field in our code.

This causes problems when the flags field is non-zero.

This previously wasn't an issue, because we did not support any command
flags.

Now I've looked at each command we support, and considered which of the
currently documented command flags we should consider. I made the
pattern matching strict for safety, so that if we encounter an unknown
future command flag that we haven't implemented, we will fail with an
error instead of ignoring it.

This is similar to what is recommended by the NBD protocol:
"The server SHOULD return EINVAL if it receives an unknown command flag.
It also SHOULD return EINVAL if it receives a request with a flag not
explicitly documented as applicable to the given request."
Except that we fail instead of returning EINVAL.
This behaviour will probably be correct most of the time, because the
protocol also says:
"Clients MUST NOT set a command flag bit that is not documented for the
particular command; and whether a flag is valid may depend on
negotiation during the handshake phase."
And if the client violates a MUST condition, the server can initiate a
hard disconnect.

Therefore this behaviour probably won't cause problems in case of
properly implemented clients, because the client must not send any
command flag that is not allowed by the transmission flags from the
server.

The following command flags are documented in the NBD protocol at the
time of writing:

- NBD_CMD_FLAG_FUA: valid for all commands, if NBD_FLAG_SEND_FUA has
  been set
- NBD_CMD_FLAG_NO_HOLE: valid for NBD_CMD_WRITE_ZEROES
- NBD_CMD_FLAG_DF: valid for NBD_CMD_READ, if NBD_FLAG_SEND_DF has been
  set
- NBD_CMD_FLAG_REQ_ONE: valid for NBD_CMD_BLOCK_STATUS

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>