ci: use a single workflow
Because publishing docs was done in its own workflow, the last ci change
missed it and made the code fail to build, because xs-opam was not ready
for this. This would have been avoided if it had been built using the
standard opam repository.

To make sure this failure doesn't happen again, integrate building a
publishing the docs in the existing workflow.

Gate only the publishing because we want to avoid repetition; and
building docs is pretty fast

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>